test(app): enforce bounded manifest replay
